@@ -1,5 +1,5 @@
 module github.com/grandecola/smaps
 
-go 1.13
+go 1.23
 
 require github.com/grandecola/mmap v0.6.0
